version1/pyproject.toml
[project]
name = "spacex-cli"
version = "0.1.0"
description = "SpaceX launch intelligence from your terminal"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Your Name", email = "you@example.com"}]
dependencies = [
    "typer[all] >=0.12, <1",
    "requests >=2.31, <3",
    "rich >=13, <14",
    "python-dotenv >=1, <2",
]

[project.optional-dependencies]
dev = [
    "pytest >=7, <8",
    "pytest-cov >=4, <5",
    "responses >=0.24, <1",
    "ruff >=0.3, <1",
]

[project.scripts]
spacex = "spacex_cli.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=spacex_cli --cov-report=term-missing --cov-fail-under=60"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
